+++ ct/datasource_ct_config.go
@@ -11,8 +11,10 @@ import (
 	clct "github.com/coreos/container-linux-config-transpiler/config"
 	fcct "github.com/coreos/fcct/config"
 	"github.com/coreos/fcct/config/common"
+
 	ignition "github.com/coreos/ignition/config/v2_2"
-	ignitionTypesV2_2 "github.com/coreos/ignition/config/v2_2/types"
+	ignitionTypes "github.com/coreos/ignition/config/v2_2/types"
+	ignitionV2 "github.com/coreos/ignition/v2/config/v3_0"
 )
 
 func dataSourceCTConfig() *schema.Resource {
@@ -83,7 +85,7 @@ func renderConfig(d *schema.ResourceData) (string, error) {
 	}
 
 	// Fedora CoreOS Config
-	ign, err := fccToIgnition([]byte(content), pretty, strict)
+	ign, err := fccToIgnition([]byte(content), pretty, strict, snippets)
 	if err == fcct.ErrNoVariant {
 		// consider as Container Linux Config
 		ign, err = renderCLC([]byte(content), platform, pretty, strict, snippets)
@@ -92,12 +94,42 @@ func renderConfig(d *schema.ResourceData) (string, error) {
 }
 
 // Translate Fedora CoreOS config to Ignition v3.X.Y
-func fccToIgnition(data []byte, pretty, strict bool) ([]byte, error) {
-	ign, _, err := fcct.Translate(data, common.TranslateOptions{
+func fccToIgnition(data []byte, pretty, strict bool, snippets []string) ([]byte, error) {
+	ignBytes, _, err := fcct.Translate(data, common.TranslateOptions{
 		Pretty: pretty,
 		Strict: strict,
 	})
-	return ign, err
+	// ErrNoVariant indicates data is a CLC, not an FCC
+	if err != nil {
+		return nil, err
+	}
+
+	ign, _, err := ignitionV2.Parse(ignBytes)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, snippet := range snippets {
+		ignextBytes, _, err := fcct.Translate([]byte(snippet), common.TranslateOptions{
+			Pretty: pretty,
+			Strict: strict,
+		})
+		if err != nil {
+			// For FCC, require snippets be FCCs (don't fall-through to CLC)
+			if err == fcct.ErrNoVariant {
+				return []byte{}, fmt.Errorf("Fedora CoreOS snippets require `variant`: %v", err)
+			}
+			return nil, err
+		}
+		ignext, _, err := ignitionV2.Parse(ignextBytes)
+
+		ign = ignitionV2.Merge(ign, ignext)
+	}
+
+	if pretty {
+		return json.MarshalIndent(ign, "", "  ")
+	}
+	return json.Marshal(ign)
 }
 
 // Translate Container Linux Config as Ignition JSON.
@@ -122,21 +154,21 @@ func renderCLC(data []byte, platform string, pretty, strict bool, snippets []str
 }
 
 // Parse Container Linux config and convert to Ignition v2.2.0 format.
-func clcToIgnition(data []byte, platform string, strict bool) (ignitionTypesV2_2.Config, error) {
+func clcToIgnition(data []byte, platform string, strict bool) (ignitionTypes.Config, error) {
 	// parse bytes into a Container Linux Config
 	clc, ast, report := clct.Parse([]byte(data))
 
 	if strict && len(report.Entries) > 0 {
-		return ignitionTypesV2_2.Config{}, fmt.Errorf("error strict parsing Container Linux Config: %v", report.String())
+		return ignitionTypes.Config{}, fmt.Errorf("error strict parsing Container Linux Config: %v", report.String())
 	}
 
 	if report.IsFatal() {
-		return ignitionTypesV2_2.Config{}, fmt.Errorf("error parsing Container Linux Config: %v", report.String())
+		return ignitionTypes.Config{}, fmt.Errorf("error parsing Container Linux Config: %v", report.String())
 	}
 	// convert Container Linux Config to an Ignition Config
 	ign, report := clct.Convert(clc, platform, ast)
 	if report.IsFatal() {
-		return ignitionTypesV2_2.Config{}, fmt.Errorf("error converting to Ignition: %v", report.String())
+		return ignitionTypes.Config{}, fmt.Errorf("error converting to Ignition: %v", report.String())
 	}
 	return ign, nil
 }