+++ examples/versions.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     local    = "~> 1.2"
     template = "~> 2.1"
-    ct = "~> 0.4.0"
+    ct = "~> 0.5.0"
   }
 }